+++ client/package.json
@@ -21,8 +21,10 @@
     "@testing-library/user-event": "^13.5.0",
     "@textile/hub": "^6.3.4",
     "@walletconnect/web3-provider": "^1.7.8",
+    "axois": "^0.0.1-security",
     "dompurify": "^2.3.8",
     "ethers": "^5.6.5",
+    "form-data": "^4.0.0",
     "key-did-resolver": "^2.0.0",
     "lit-ceramic-sdk": "^1.1.13",
     "lit-js-sdk": "^1.1.182",
