+++ client/src/components/Upload/index.js
@@ -1,0 +1,1 @@
+export * from './Upload.jsx';
