--- client/src/components/write/index.js
@@ -1,1 +1,0 @@
-export * from './Write.jsx';
